+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.3.1
+
+  - operate on runtime types instead of declared types whenever possible 
+
 ## v0.3.0
 
   - iterables tab now provides additional operations "count" and "group"